wx v2.7.0 incremental update - minor
git-svn-id: http://svn.osafoundation.org/chandler/trunk@9946 3c49585b-f0f7-0310-b5f9-dfe92a88fbfe

+++ internal/wx/src/gtk/radiobox.cpp
@@ -2,7 +2,7 @@
 // Name:        src/gtk/radiobox.cpp
 // Purpose:
 // Author:      Robert Roebling
-// Id:          $Id: radiobox.cpp,v 1.116 2006/03/23 22:04:56 VZ Exp $
+// Id:          $Id: radiobox.cpp,v 1.117 2006/03/24 08:31:58 ABX Exp $
 // Copyright:   (c) 1998 Robert Roebling
 // Licence:     wxWindows licence
 /////////////////////////////////////////////////////////////////////////////
@@ -226,7 +226,7 @@ bool wxRadioBox::Create( wxWindow *parent, wxWindowID id, const wxString& title,
 
     wxString label;
     GSList *radio_button_group = (GSList *) NULL;
-    for (unsigned int i = 0; i < n; i++)
+    for (unsigned int i = 0; i < (unsigned int)n; i++)
     {
         if ( i != 0 )
             radio_button_group = gtk_radio_button_get_group( GTK_RADIO_BUTTON(m_radio) );
